Fix additional variable type errors

+++ Practice.cpp
@@ -42,7 +42,7 @@ bool Practice::isPalindrome(string input)
   }
   for(int i=0; i < (int)input.size()/2; i++)
   {
-    if( input[i] != input[input.size()-1-i] )
+    if( input[i] != input[(int)input.size()-1-i] )
       return false;
   }
   return true;
